### 2024-01-23 #6273 @editorialbot
[PRE REVIEW]: Jury: A Comprehensive Evaluation Toolkit
**Submitting author:** @devrimcavusoglu (<a href="http://orcid.org/0000-0002-5218-1283">Devrim Çavuşoğlu</a>)
**Repository:** https://github.com/obss/jury
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.3
