### 2024-01-23 #6273 @arfon
@devrimcavusoglu  – thanks for your submission to JOSS. In your submission you said:

> A detailed paper describing the framework and the library submitted to three ACL venues (e.g. system demonstration tracks of ACL'23, EMNLP'23, EACL'24) earlier.

Could you please share a link to these papers (or upload them here).
