### 2024-01-23 #6273 @devrimcavusoglu
@arfon Of course, I thought I also provided the paper in the note. [Link](https://arxiv.org/abs/2310.02040) to the paper.
