### 2024-01-23 #6273 @arfon
> @arfon Of course, I thought I also provided the paper in the note. [Link](https://arxiv.org/abs/2310.02040) to the paper.

Ah, sorry. This _is_ the paper that has been submitted to those venues? If so, could you tell me more about what ACL'23, EMNLP'23, EACL'24 are? Are those submissions reviewed and published conference papers?
