### 2024-01-23 #6273 @devrimcavusoglu
@arfon Yes, it is. All the submission/review phases are completed, the paper has not been published on those venues, and there will be no further other submissions.
