### 2024-01-31 #6273 @arfon
Thanks for the additional background @devrimcavusoglu. My concern here is that we might (as a journal) consider this a repeat publication for the same work which we don't allow. I'm going to ask the wider editorial team for their input here.
