### 2024-01-31 #6273 @devrimcavusoglu
@arfon Thanks for the response, but arxiv is only an online preprint and it's not considered as a publication as in many venues (including journals, conferences, etc.) publishing the paper as an online preprint does not invalidate submissions to the venues. Having said that I don't throughly know your stance on this, so we'll be waiting for your response.
