### 2024-02-01 #6273 @devrimcavusoglu
> @devrimcavusoglu - I'm another track editor looking at this. Can you help me understand how I should interpret
>
> > A detailed paper describing the framework and the library submitted to three ACL venues (e.g. system demonstration tracks of ACL'23, EMNLP'23, EACL'24) earlier.
>
> and
>
> > All the submission/review phases are completed, the paper has not been published on those venues, and there will be no further other submissions.
>
> I don't think I understand the model of these conferences, if you submit work to them and they then don't publish it, unless you mean that the submissions to these conferences were not accepted?

Of course, I mean I thought that was obvious enough and didn't mention it explicitly. The paper was refactored w.r.t. all comments and published on arxiv by us, marking the current version of the paper.
