### 2024-02-01 #6273 @danielskatz
Thanks! This wasn't obvious to me. It was unclear what "submitted" meant, given that the work could still be under consideration. This also clarifies the lack of any publication duplication concern, in my opinion.
