### 2024-02-17 #6273 @arfon
Thanks for the feedback @devrimcavusoglu. I've removed the scope review label, but now we're in a 'waitlisted' state. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
