### 2024-03-08 #6452 @KennethEnevoldsen
@devrimcavusoglu this looks like a wonderful work, which I am sure will be beneficial to many people (seems like it already is). Overall I was able to run both the examples, tests and installation without issues and found that almost all of the information (required and desired) was available.

I have a few things I was hoping you could clear up:

- it seems like there are no guidelines for how to report issues or seek support. I would add those in (it can simply be a table e.g. as seen [here](https://github.com/centre-for-humanities-computing/DaCy).
- Have the metrics been evaluated against the implementation of e.g. Huggingface metrics (where meaningful)?
