### 2024-03-11 #6452 @evamaxfield
@devrimcavusoglu seconding that this is really great and I will likely start using it in some of my own work :)

Some comments:

1. (**Not required but I tend to think its nice**) I tried BERTScore in addition to the defaults and it asked me to install bert-score which I completely understand not including as a default install. I generally like it when packages provide an extra install for the required extra dependency(ies). For example `pip install jury[bertscore]`. It also allows you to put version pins on the extra dependencies to make sure that the version also works with your software.
2. Also seconding that I didn't see any community guidelines or other documentation specifically around how to contribute or ways to contribute. Please add those.
3. I am fine with not having a full documentation page or website but I do think it would be nice to have. At the very least can you include a link to the examples notebook somewhere in the README as "See more examples here" or something. (Although I guess this might be a question for @crvernon -- there is API docs in the code but to my knowledge I didn't see a documentation website where those API docs can be easily seen, is this an issue? I am fine with it because I am fine with digging through the source code but for others / general dev experience it may be a nice to add.)
