### 2024-03-13 #6452 @crvernon
@evamaxfield to your question:

> I am fine with not having a full documentation page or website but I do think it would be nice to have. At the very least can you include a link to the examples notebook somewhere in the README as "See more examples here" or something. (Although I guess this might be a question for @crvernon -- there is API docs in the code but to my knowledge I didn't see a documentation website where those API docs can be easily seen, is this an issue? I am fine with it because I am fine with digging through the source code but for others / general dev experience it may be a nice to add.)

We don't require documentation to be hosted separately; though it is a nice complement that many reviewers have started to point out.  I would say this falls under the category of suggestion but not requirement as long as the codebase has the necessary documentation that is visible to the user.  Thanks!
