### 2024-03-13 #6452 @evamaxfield
Sounds good! Then my comment resolves to simply adding a link to the examples notebook.
