### 2024-03-20 #6452 @devrimcavusoglu
@KennethEnevoldsen @evamaxfield Thank you for the reviews and feedback. Apologizes for a late response. I will try to respond to items that you mentioned.

**Response to @KennethEnevoldsen:**
> * it seems like there are no guidelines for how to report issues or seek support. I would add those in (it can simply be a table e.g. as seen [here](https://github.com/centre-for-humanities-computing/DaCy).

We did not actually documented this as we think for most of the open-source software the Github platform is the de facto standard for raising issues, bug reports, discussions and etc. If we were using another platform for those stuff we could've mentioned it which is reasonable; however, utilizing the standard does not necessitate this as it is the standard and easy to be grasped by all users. Nonetheless, we particularly included templates for the issues (e.g. bug reports, feature/metric request).

> * Have the metrics been evaluated against the implementation of e.g. Huggingface metrics (where meaningful)?

Yes, there is a comparison in [the paper](https://arxiv.org/pdf/2310.02040.pdf), particularly Fig. 2. Jury is compared to HF metrics package in two different perspectives, input size and number of metrics.

**Response to @evamaxfield:**

> Sounds good! Then my comment resolves to simply adding a link to the examples notebook.

Actually, we do have link to the notebook as a Google Colab link although it's not expressed and written down, for example, in a table in README. This way we think is more user friendly as the users can directly engage with the package and play around.

![image](https://github.com/openjournals/joss-reviews/assets/46989091/9402d90b-6742-4c03-9318-1bce203f419e)

> (Not required but I tend to think its nice) I tried BERTScore in addition to the defaults and it asked me to install bert-score which I completely understand not including as a default install. I generally like it when packages provide an extra install for the required extra dependency(ies). For example pip install jury[bertscore]. It also allows you to put version pins on the extra dependencies to make sure that the version also works with your software.

That is a really good suggestion, we may definitely plan to add this feature in upcoming versions.
