### 2024-03-20 #6452 @KennethEnevoldsen
> We did not actually documented this as we think for most of the open-source software the Github platform is the de facto standard for raising issues, bug reports, discussions and etc. If we were using another platform for those stuff we could've mentioned it which is reasonable; however, utilizing the standard does not necessitate this as it is the standard and easy to be grasped by all users. Nonetheless, we particularly included templates for the issues (e.g. bug reports, feature/metric request).

I am perfectly fine with this as long as the editors (@crvernon) agree. I will leave it unchecked on the checklist.

> Yes, there is a comparison in [the paper](https://arxiv.org/pdf/2310.02040.pdf), particularly Fig. 2. Jury is compared to HF metrics package in two different perspectives, input size and number of metrics.

Thanks for the clarification

I believe this clarifies all my points.
