### 2024-03-22 #6452 @crvernon
Per @KennethEnevoldsen 's comment...

> > We did not actually documented this as we think for most of the open-source software the Github platform is the de facto standard for raising issues, bug reports, discussions and etc. If we were using another platform for those stuff we could've mentioned it which is reasonable; however, utilizing the standard does not necessitate this as it is the standard and easy to be grasped by all users. Nonetheless, we particularly included templates for the issues (e.g. bug reports, feature/metric request).
>
> I am perfectly fine with this as long as the editors (@crvernon) agree. I will leave it unchecked on the checklist.

I do believe there needs to be a bit more here to meet the guidelines for:

>  **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Even though I do agree that many users may be aware of what issues and pull requests are, our papers go out to a general audience.  Thus, questions and contributions may not always come in the form of code modifications/additions from engineers, but from program managers, etc. who seek clarity on the purpose of the software in the context of their mission.  These audiences often do not have familiarity with how GitHub works.  So I do advise @devrimcavusoglu to be more descriptive for how to contribute.  Feel free to check out previous JOSS publications for some examples.  Thanks!
