### 2024-03-22 #6452 @devrimcavusoglu
> Even though I do agree that many users may be aware of what issues and pull requests are, our papers go out to a general audience. Thus, questions and contributions may not always come in the form of code modifications/additions from engineers, but from program managers, etc. who seek clarity on the purpose of the software in the context of their mission. These audiences often do not have familiarity with how GitHub works. So I do advise @devrimcavusoglu to be more descriptive for how to contribute. Feel free to check out previous JOSS publications for some examples. Thanks!

@crvernon I see the point, thanks for clarfying. I will make appropriate updates in the repo ASAP and will ping you then.
