### 2024-04-03 #6452 @KennethEnevoldsen
With the addition of community guidelines I don't believe there is any additions I want to add. I believe it is a good paper/package which provides solid tools for developers.
