### 2024-04-05 #6452 @devrimcavusoglu
@crvernon I updated the repo for regarding changes. You can access to [the community guidline](https://github.com/obss/jury?tab=readme-ov-file#community-interaction) in the repo.

Thank you for review and fruitful discussion. @KennethEnevoldsen @evamaxfield @crvernon
