### 2024-04-05 #6452 @crvernon
:wave: @devrimcavusoglu - Great work on this!  I just need you to address a few minor things in the paper and we can move forward with next steps.

**Paper**:

- [ ] **LINE 57-58**:  The following reads odd, "Our library also exploits `datasets` package to promote the open-source contribution..." perhaps change this to "Our library also utilizes the `datasets` package to promote open-source contribution..."
- [ ] **LINE 59-60**: "...released by `datasets` package..." should be "...released by the `datasets` package..."
