+++ paper.md
@@ -38,18 +38,32 @@ Evaluation plays a critical role in deep learning as a fundamental block of any 
 
 # Statement of need
 
-NLP tasks possess inherent complexity, requiring a comprehensive evaluation of model performance beyond a single metric comparison. Established benchmarks such as WMT [@barrault-etal-2020-findings] and GLUE [@wang-etal-2018-glue] rely on multiple metrics to evaluate models on standardized datasets. This practice promotes fair comparisons across different models and pushes advancements in the field. Embracing multiple metric evaluations provides valuable insights into a model's generalization capabilities. By considering diverse metrics, such as accuracy, F1 score, BLEU, and ROUGE, researchers gain a holistic understanding of a model's response to never-seen inputs and its ability to generalize effectively. Furthermore, task-specific NLP metrics enable the assessment of additional dimensions, such as readability, fluency, and coherence. The comprehensive evaluation facilitated by multiple metric analysis allows for trade-off studies and aids in assessing generalization for task-independent models. Given these numerous advantages, NLP specialists lean towards employing multiple metric evaluations. 
+NLP tasks possess inherent complexity, requiring a comprehensive evaluation of model performance beyond a single metric comparison. Established benchmarks such as WMT **[1]** and GLUE **[2]** rely on multiple metrics to evaluate models on standardized datasets. This practice promotes fair comparisons across different models and pushes advancements in the field. Embracing multiple metric evaluations provides valuable insights into a model's generalization capabilities. By considering diverse metrics, such as accuracy, F1 score, BLEU, and ROUGE, researchers gain a holistic understanding of a model's response to never-seen inputs and its ability to generalize effectively. Furthermore, task-specific NLP metrics enable the assessment of additional dimensions, such as readability, fluency, and coherence. The comprehensive evaluation facilitated by multiple metric analysis allows for trade-off studies and aids in assessing generalization for task-independent models. Given these numerous advantages, NLP specialists lean towards employing multiple metric evaluations. 
 
 Although employing multiple metric evaluation is common, there is a challenge in practical use because widely-used metric libraries lack support for combined and/or concurrent metric computations. Consequently, researchers face the burden of evaluating their models per metric, a process exacerbated by the scale and complexity of recent models and limited hardware capabilities. This bottleneck impedes the efficient assessment of NLP models and highlights the need for enhanced tooling in the metric computation for convenient evaluation. In order for concurrency to be beneficial at a maximum level, the system may require hardware accordingly. Having said that, the availability of the hardware comes into question. 
 
-The extent of achievable concurrency in NLP research has traditionally relied upon the availability of hardware resources accessible to researchers. However, significant advancements have occurred in recent years, resulting in a notable reduction in the cost of high-end hardware, including multi-core CPUs and GPUs. This progress has transformed high-performance computing resources, which were once prohibitively expensive and predominantly confined to specific institutions or research labs, into more accessible and affordable assets. For instance, in BERT [@devlin-etal-2019-bert] and XLNet [@zhilin-etal-2019-xlnet], it is stated that they leveraged the training process by using powerful yet cost-effective hardware resources. Those advancements show that the previously constraining factor for hardware accessibility has been mitigated, allowing researchers to overcome the limitations associated with achieving concurrent processing capabilities in NLP research.
[…]
