### 2024-04-06 #6452 @crvernon
@devrimcavusoglu - Thanks!  However, there was no need to refactor the references.  The previous version was how we prefer them to be used in JOSS.  Please revert the citation and reference format back to the way they were.

Also, this phrase in the paper is still incorrect:

**LINES 56-58**

Current version:
> Our library also utilizes `datasets` package to promote the open-source contribution; when users implement metrics, the implementation can be contributed to the `datasets` package.

How the sentence should read:
> Our library also utilizes the `datasets` package to promote open-source contribution; when users implement metrics, the implementation can be contributed to the `datasets` package.

Please let me know when these changes have been made.
