### 2024-04-20 #6452 @crvernon
This is looking great @devrimcavusoglu!  Just one more minor thing to address in the paper before we move on...

On LINE 80, "Bleu" should appear in all caps as "BLEU".  You can make this happen by bracketing (e.g., {BLEU}) the letters you wish to remain capitalized in the bib file.

Let me know when this has been updated and we will move on to the next step.  Thanks!
