### 2024-04-20 #6452 @devrimcavusoglu
> On LINE 80, "Bleu" should appear in all caps as "BLEU". You can make this happen by bracketing (e.g., {BLEU}) the letters you wish to remain capitalized in the bib file.

This is intended, I didn't want to modify the bibtex citation given by ACL, refer to [the publication page](https://aclanthology.org/P02-1040/) for citations. I used the bib citation as is.
