### 2024-05-10 #6452 @devrimcavusoglu
Hi @crvernon, thanks for the follow-up, actually `jury` already has a [Zenodo archive](https://zenodo.org/records/6109838). I can update a new version with the latest changes, I'm not sure I should do so on the `joss-paper` branch including the JOSS paper or in `main` branch. Can you please clarify if new version is required (if not you can just go with the available zenodo page) ?
