+++ jury/__init__.py
@@ -1,4 +1,4 @@
 from jury.core import Jury
 from jury.metrics import AutoMetric, list_metrics, load_metric
 
-__version__ = "2.3"
+__version__ = "2.3.1"
