### 2024-05-20 #6452 @crvernon
@devrimcavusoglu the titles still do not match.

- [ ] Please fix the title
**What is in your paper**:   Jury: A Comprehensive Evaluation Toolkit
**The title of your archive**:  Jury: A Comprehensive NLP Evaluation toolkit

 - [ ] Also, the author order is still incorrect in your archive as compared to what it is in the paper.

**Please fix both issue to match what is in the paper.**
