### 2024-05-20 #6452 @crvernon
@devrimcavusoglu - We have no problems with the preprint.  See https://joss.readthedocs.io/en/latest/submitting.html#preprint-policy

Just let me know if you alter the paper and I'll recheck.  If not, just let me know and we can move forward.  Thanks!
