### 2024-05-20 #6452 @crvernon
🥳 Congratulations on your new publication @devrimcavusoglu! Many thanks to @evamaxfield and @KennethEnevoldsen for your time, hard work, and expertise in reviewing!! JOSS wouldn't be able to function nor succeed without your efforts.
